# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Malmö Resturant Api",
        "version": "0.1"
    },
    "paths": {
...
        "/api/itemimages/new": {
            "post": {
                "tags": [
                    "Item Images"
                ],
                "description": "Create a new item image",
                "operationId": "App\\Http\\Controllers\\ItemImagesController::create",
                "parameters": [
                    {
                        "name": "image",
                        "in": "query",
                        "description": "Url to an image",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "file"
                        }
                    },
                    {
                        "name": "item_id",
                        "in": "query",
                        "description": "item_id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "file"
                        }
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Api Token",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "file"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Create a new item image"
                    }
                }
            }
        },
        "/api/itemimages/{id}": {
            "get": {
                "tags": [
                    "Item Images"
                ],
                "description": "Get a specific item image",
                "operationId": "App\\Http\\Controllers\\ItemImagesController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Id of item image",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "file"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Get a specific item image"
                    }
                }
            }
        },
        "/api/itemimages/report/{id}": {
            "put": {
                "tags": [
                    "Item Images"
                ],
                "description": "Report an item image",
                "operationId": "App\\Http\\Controllers\\ItemImagesController::report",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Id of item image",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "file"
                        }
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Api Token",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "file"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Report an item review"
                    }
                }
            }
        },
        "/api/itemimages/unreport/{id}": {
            "put": {
                "tags": [
                    "Item Images"
                ],
                "description": "unreport an item review",
                "operationId": "App\\Http\\Controllers\\ItemImagesController::unreport",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Id of item reviews",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "file"
                        }
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Api Token",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "file"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "unreport an item review"
                    }
                }
            }
        },
...
        "/api/itemreviews/unreport/{id}": {
            "put": {
                "tags": [
                    "Item Reviews"
                ],
                "description": "unreport an item review",
                "operationId": "App\\Http\\Controllers\\ItemReviewsController::unreport",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Id of item reviews",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "file"
                        }
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Api Token",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "file"
                        }
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Api Token",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "file"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "unreport an item review"
                    }
                }
            }
        },
...
        "/api/items/delete/{id}": {
            "delete": {
                "tags": [
                    "Items"
                ],
                "description": "Delete a specific item",
                "operationId": "App\\Http\\Controllers\\ItemsController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Id of item",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "file"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Delete a specific item"
                    }
                }
            }
        },
...
        "/api/items/unreport/{id}": {
            "put": {
                "tags": [
                    "Items"
                ],
                "description": "unreport an item",
                "operationId": "App\\Http\\Controllers\\ItemsController::unreport",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Id of item",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "file"
                        }
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Api Token",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "file"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "unreport an item"
                    }
                }
            }
        },
...
        "/api/resturantimages/unreport/{id}": {
            "put": {
                "tags": [
                    "Resturant Images"
                ],
                "description": "unreport an item review",
                "operationId": "App\\Http\\Controllers\\ResturantImagesController::unreport",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Id of item reviews",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "file"
                        }
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Api Token",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "file"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "unreport an item review"
                    }
                }
            }
        },
...
        "/api/resturantreviews/unreport/{id}": {
            "put": {
                "tags": [
                    "Resturant Reviews"
                ],
                "description": "unreport an Resturant Review",
                "operationId": "App\\Http\\Controllers\\ResturantReviewsController::unreport",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Id of item",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "file"
                        }
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Api Token",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "file"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "unreport an Resturant Review"
                    }
                }
            }
        },
...
        "/api/resturants/unreport/{id}": {
            "put": {
                "tags": [
                    "Resturants"
                ],
                "description": "unreport a Resturant",
                "operationId": "App\\Http\\Controllers\\ResturantsController::unreport",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Id of resturant",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "file"
                        }
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Api Token",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "file"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "unreport a Resturant"
                    }
                }
            }
        }
    }
}